% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lowestoft_stuff.R
\name{read_lowestoft}
\alias{read_lowestoft}
\title{Read in Lowestoft-format VPA data}
\usage{
read_lowestoft(filename, val.name, format = "List")
}
\arguments{
\item{filename}{The name of the file to read in}

\item{val.name}{XXX}

\item{format}{The format of the output, available is "List","Wide","Long"}
}
\description{
Read in Lowestoft-format VPA data
}

