% Generated by roxygen2 (4.0.1): do not edit by hand
\name{read_lowestoft_survey}
\alias{read_lowestoft_survey}
\title{Read Lowestoft-format survey data}
\usage{
read_lowestoft_survey(filename, format = "long")
}
\arguments{
\item{filename}{The name of the file to read in}

\item{format}{XXX}
}
\description{
Read Lowestoft-format survey data
}

