% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_hake.R
\name{read_rbx_hake}
\alias{read_rbx_hake}
\title{read_result_hake}
\usage{
read_rbx_hake(infilename, outfilename, runname)
}
\arguments{
\item{infilename}{A character string incuding the path (directory of the input file name.
For hake the name of the file is normally namhakdata.dat}

\item{outfilename}{A character string, including path (directory) to the report
file. The report file has normally the ending .rep.}

\item{runname}{A character specifying the name of the run}
}
\description{
Reads the ADMB report (result) file from the Namibian hake assessment
}

