% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_horsemackerel.R
\name{read_rbx_hmac}
\alias{read_rbx_hmac}
\title{Reads the ADMB report (result) file from the Namibian horse mackerel assessment}
\usage{
read_rbx_hmac(filename)
}
\arguments{
\item{filename}{A character string, including path (directory) to the report
file. The report file has normally the ending .rep.}
}
\description{
XXX
}

