% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sam_stuff.R
\name{read_rbya_sam}
\alias{read_rbya_sam}
\title{read_rbya_sam}
\usage{
read_rbya_sam(x, Scale = 1)
}
\arguments{
\item{x}{Object from read_fit_sam}

\item{Scale}{A value}
}
\description{
read fishing mortality and stock in numbers from sam
}

