% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_hake.R
\name{read_sim_hake}
\alias{read_sim_hake}
\title{read_sim_hake}
\usage{
read_sim_hake(path)
}
\arguments{
\item{path}{The directory path where the Files are located files are located}
}
\value{
Returns a list containing two data.frames, the raw data (raw) and
summarized data (qua). The latter contains the 0.05, 0.25, 0.50, 0.75 and
0.95 percentiles as well as the mean.
}
\description{
Reads the simulation files for Hake.
}

