% Generated by roxygen2 (4.0.1): do not edit by hand
\name{str_clean}
\alias{str_clean}
\title{str_clean}
\usage{
str_clean(txt, p1, p2, Replace = "", header = FALSE, sep = "", ...)
}
\arguments{
\item{txt}{A text vector}

\item{p1}{Numerical value specifying 1st position}

\item{p2}{Numerical value specifying last position}

\item{Replace}{A text string}

\item{header}{Boolean}

\item{sep}{A text string}

\item{...}{Additional arguments}
}
\description{
A helper function
}

