% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidy_functions.R
\name{str_trim_all}
\alias{str_trim_all}
\title{Trims whitespace from both sides and extra spaces within string}
\usage{
str_trim_all(string)
}
\arguments{
\item{string}{A character vector}
}
\description{
XXX
}

