% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidy_functions.R
\name{str_trim_tab}
\alias{str_trim_tab}
\title{str_trim_tab}
\usage{
str_trim_tab(txt, side = "both")
}
\arguments{
\item{txt}{input character vector}

\item{side}{side on which tab is removed (left, right, both)}
}
\description{
Trim tabs from start, end or both sides of string
}

