% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/x_per_recruit_stuff.R
\name{ypr}
\alias{ypr}
\title{yield per recruit}
\usage{
ypr(df, fscan = 0, a1, a2, plusgroup = TRUE, recruitment = 1, pM = 0,
  pF = 0)
}
\arguments{
\item{df}{\code{data.frame}}

\item{fscan}{A numerical vector specifying fishing mortality to scan over}

\item{a1}{Value containing lower reference age}

\item{a2}{Value containing upper reference age}

\item{plusgroup}{A boolean. If TRUE (default) the last age group is treated
as a plus group.}

\item{recruitment}{An number specifying number of recruits. If one (default)
returns the conventional yield and ssb for one recruit.}

\item{pM}{Value containing the proportion of M prior to spawning}

\item{pF}{Value containing the proportion of F prior to spawning}
}
\value{
returns \code{list} with catch and ssb for each age
}
\description{
Calculates yield and ssb per recruit. 
Summarization over age is provided by the function \code{ypra}
}

