% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/x_per_recruit_stuff.R
\name{ypra}
\alias{ypra}
\title{yield per recruit by age}
\usage{
ypra(x, fmort = 0, a1, a2, plusgroup = TRUE, recruitment = 1, pM = 0,
  pF = 0)
}
\arguments{
\item{x}{\code{data.frame}}

\item{fmort}{A numerical value specifying fishing mortality}

\item{a1}{Value containing lower reference age}

\item{a2}{Value containing upper reference age}

\item{plusgroup}{A boolean. If TRUE (default) the last age group is treated
as a plus group.}

\item{recruitment}{An number specifying number of recruits. If one (default)
returns the conventional yield and ssb for one recruit.}

\item{pM}{Value containing the proportion of M prior to spawning}

\item{pF}{Value containing the proportion of F prior to spawning}
}
\value{
returns \code{list} with catch and ssb for each age
}
\description{
Returns a detailed (by age) yield and ssb composition. 
Summarization over age is provided by the function \code{ypr}
}

