% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sam_stuff.R
\name{read_conf_sam}
\alias{read_conf_sam}
\title{read_conf_sam}
\usage{
read_conf_sam(filename)
}
\arguments{
\item{filename}{A character vector containg path and name to ....}
}
\description{
Reads sam configuration file and returns a list. 
Used by read_rby_sam, xxx.
This function is identical(?) to originally in scr/common.R
}
\author{
Anders Nielsen
}

