% Generated by roxygen2 (4.0.1): do not edit by hand
\name{read_input_hake}
\alias{read_input_hake}
\title{read_input_hake}
\usage{
read_input_hake(filename)
}
\arguments{
\item{filename}{A character string, including path (directory) to the report
file. The report file has normally the ending .rep.}
}
\description{
Reads the ADMB input data file from the Namibian hake assessment
}

