% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sam_stuff.R
\name{read_rby_sam}
\alias{read_rby_sam}
\title{read_rby_sam}
\usage{
read_rby_sam(conf_sam, format = "long")
}
\arguments{
\item{conf_sam}{An list object returned by function \code{read_conf_sam}}

\item{format}{A character vector specifying output format. If default ("long") then function
returns a data.frame ....}
}
\description{
Read sam results by year and age
}

